+++ YoutubeDownloader/Converters/VideoQualityPreferenceToStringConverter.cs
@@ -10,7 +10,7 @@ namespace YoutubeDownloader.Converters
     {
         public static VideoQualityPreferenceToStringConverter Instance { get; } = new();
 
-        public object? Convert(object? value, Type targetType, object parameter, CultureInfo culture)
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
             if (value is VideoQualityPreference preference)
                 return preference.GetDisplayName();
@@ -18,7 +18,7 @@ namespace YoutubeDownloader.Converters
             return default(string);
         }
 
-        public object ConvertBack(object? value, Type targetType, object parameter, CultureInfo culture) =>
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) =>
             throw new NotSupportedException();
     }
 }