+++ YoutubeDownloader/Converters/VideoToLowestQualityThumbnailUrlConverter.cs
@@ -11,7 +11,7 @@ namespace YoutubeDownloader.Converters
     {
         public static VideoToLowestQualityThumbnailUrlConverter Instance { get; } = new();
 
-        public object? Convert(object? value, Type targetType, object parameter, CultureInfo culture)
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
             if (value is IVideo video)
                 return video.Thumbnails.OrderBy(t => t.Resolution.Area).FirstOrDefault()?.Url;
@@ -19,7 +19,7 @@ namespace YoutubeDownloader.Converters
             return default(string);
         }
 
-        public object ConvertBack(object? value, Type targetType, object parameter, CultureInfo culture) =>
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) =>
             throw new NotSupportedException();
     }
 }