+++ YoutubeDownloader/Models/VideoDownloadOption.cs
@@ -50,7 +50,7 @@ namespace YoutubeDownloader.Models
                 StringComparer.OrdinalIgnoreCase.Equals(Label, other.Label);
         }
 
-        public override bool Equals(object? obj)
+        public override bool Equals(object obj)
         {
             if (ReferenceEquals(null, obj)) return false;
             if (ReferenceEquals(this, obj)) return true;