+++ YoutubeDownloader/ViewModels/Dialogs/MessageBoxViewModel.cs
@@ -4,9 +4,8 @@ namespace YoutubeDownloader.ViewModels.Dialogs
 {
     public class MessageBoxViewModel : DialogScreen
     {
-        public string? Title { get; set; }
-
-        public string? Message { get; set; }
+        public string Title { get; set; }
+        public string Message { get; set; }
     }
 
     public static class MessageBoxViewModelExtensions